refactor licenseInfo service (together with @alex-evo)
