+++ backend/src/src-licenseinfo/pom.xml
@@ -21,6 +21,16 @@
             <artifactId>src-attachments</artifactId>
             <version>1.4.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.spdx</groupId>
+            <artifactId>spdx-tools</artifactId>
+            <version>2.0.4</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.velocity</groupId>
+            <artifactId>velocity</artifactId>
+            <version>1.6.2</version>
+        </dependency>
     </dependencies>
     <parent>
         <groupId>com.siemens.sw360</groupId>
@@ -35,6 +45,9 @@
     <developers>
         <developer>
             <name>alex.borodin@evosoft.com</name>
+        </developer>
+        <developer>
+            <name>maximilian.huber@tngtech.com</name>
         </developer>
     </developers>
 </project>
