+++ backend/src/src-licenseinfo/src/main/java/com/siemens/sw360/licenseinfo/UncheckedTException.java
@@ -1,4 +1,4 @@
-package com.siemens.sw360.datahandler.licenseinfo;
+package com.siemens.sw360.licenseinfo;
 
 import org.apache.thrift.TException;
