+++ backend/src/src-licenseinfo/src/main/resources/licenseInfoFile.vm
