+++ backend/src/src-licenseinfo/src/test/java/com/siemens/sw360/licenseinfo/parsers/CLIParserTest.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package com.siemens.sw360.datahandler.licenseinfo.parsers;
+package com.siemens.sw360.licenseinfo.parsers;
 
 import com.siemens.sw360.datahandler.couchdb.AttachmentConnector;
 import com.siemens.sw360.datahandler.thrift.attachments.Attachment;
